+++ CHANGELOG.md
@@ -9,11 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Debug logging for port filter resolution, URL building, and instance filtering in `ConsulProxyServer`
+*
 
 ### Changed
 
 *
+
+### Fixed
+
+*
+
+## [1.31.1] - 2026-03-31
+
+### Added
+
+* Debug logging for port filter resolution, URL building, and instance filtering in `ConsulProxyServer`
 
 ### Fixed